+++ lib/src/material/material_progress_bar.dart
@@ -1,3 +1,4 @@
+import 'package:chewie/src/cast/chewie_playback_target.dart';
 import 'package:chewie/src/chewie_progress_colors.dart';
 import 'package:chewie/src/progress_bar.dart';
 import 'package:flutter/material.dart';
@@ -15,12 +16,17 @@ class MaterialVideoProgressBar extends StatelessWidget {
     this.onDragUpdate,
     super.key,
     this.draggableProgressBar = true,
+    this.playback,
   }) : colors = colors ?? ChewieProgressColors();
 
   final double height;
   final double barHeight;
   final double handleHeight;
   final VideoPlayerController controller;
+
+  /// What to read and seek. Defaults to [controller]; Chewie passes the cast
+  /// receiver here while a session is live.
+  final ChewiePlaybackTarget? playback;
   final ChewieProgressColors colors;
   final Function()? onDragStart;
   final Function()? onDragEnd;
@@ -39,6 +45,7 @@ class MaterialVideoProgressBar extends StatelessWidget {
       onDragStart: onDragStart,
       onDragUpdate: onDragUpdate,
       draggableProgressBar: draggableProgressBar,
+      playback: playback,
     );
   }
 }